backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
SQLAlchemy==2.0.29
python-dotenv==1.0.1
pydantic-settings==2.2.1
alembic==1.13.1
psycopg[binary]==3.1.18
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
python-multipart==0.0.9
bcrypt==4.0.1
